% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesRange-package.R
\docType{data}
\name{Euchloe_ausonia}
\alias{Euchloe_ausonia}
\title{Euchloe ausonia points exemplary file}
\description{
A \code{DataFrame} species occurrence
}
\examples{

data("Euchloe_ausonia")

}
\keyword{exemplary}
\keyword{object}
