% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesRange-package.R
\docType{data}
\name{Euchloe_ausonia_shp}
\alias{Euchloe_ausonia_shp}
\title{Euchloe ausonia range map}
\description{
A \code{SpatialPolygonsDataFrame} convex hull
}
\examples{

data("Euchloe_ausonia_shp")

}
\keyword{exemplary}
\keyword{object}
