% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updataRange.R
\name{updateRange}
\alias{updateRange}
\title{updateRange}
\usage{
updateRange(
  occ,
  range,
  res = 1,
  range_bonus = 75,
  occ_bonus = 20,
  abs_pen = 40,
  red2 = 0.5
)
}
\arguments{
\item{occ}{dataFrame of the species occurrence. Output of function occData.}

\item{range}{shapefile, the user can provide a specialist range map, or a
convex hull to be visualised with the point occurrence.}

\item{res}{numeric, spatial resolution in degrees. Default is 1.}

\item{range_bonus}{numeric, percentage of bonus to be given to cells within
the basis range.}

\item{occ_bonus}{numeric, value of bonus given by presences.}

\item{red2}{numeric, distance decay for \emph{occ_bunus} and \emph{abs_penalty}.}

\item{abs_penalty}{numeric, value of penalty given by absences.}
}
\value{
An updated range map based on presences and absences.
}
\description{
Update the species range map based on point-occurrence presence and absence
data. The algorithm uses either a specialist range map or a convex hull as
basis for the calculations
}
\examples{

}
