% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualiseResults.R
\name{visualiseResults}
\alias{visualiseResults}
\title{plotData}
\usage{
visualiseResults(occ, regional = TRUE, range, updated_range)
}
\arguments{
\item{occ}{dataFrame of the species occurrence. Output of function occData.}

\item{regional}{logical, whether the whole world should be plotted as the
background or only the region adjacent to the species countries of
occurrence.}

\item{range}{shapefile, the user can provide a specialist range map, or a
convex hull to be visualised with the point occurrence.}
}
\value{
This function plots the species occurrence
}
\description{
Plot the species occurrences with map background for visualisation
}
